Revert "Trips (#6)"
This reverts commit 573ac86f972d6e2e8825cd03235d7f6de1124d8c.
