--- components/datePicker/index.js
@@ -1,22 +1,0 @@
-import ReactDatePicker from 'react-datepicker';
-import 'react-datepicker/dist/react-datepicker.css';
-
-const DatePicker = ({
-  selectedDate,
-  onChange,
-  isClearable = false,
-  showPopperArrow = false,
-  ...props
-}) => {
-  return (
-    <ReactDatePicker
-      selected={selectedDate}
-      onChange={onChange}
-      isClearable={isClearable}
-      showPopperArrow={showPopperArrow}
-      {...props}
-    />
-  );
-};
-
-export default DatePicker;
