+++ components/emoji.js
@@ -1,29 +1,12 @@
-const Emoji = (props) => {
-  const getSymbol = (name) => {
-    switch (name) {
-      case "transport":
-        return "🚌";
-      case "lodging":
-        return "🏠";
-      case "travel":
-        return "✈️";
-      case "meal":
-        return "🍕";
-      case "event":
-        return "🎉";
-      default:
-        return "😀";
-    }
-  };
-  return (
-    <span
-      role="img"
-      aria-label={props.label ? props.label : ""}
-      aria-hidden={props.label ? "false" : "true"}
-      style={{ fontSize: "20px" }}
-    >
-      {(props.name && getSymbol(props.name)) || props.symbol}
-    </span>
-  );
-};
+import React from "react";
+const Emoji = (props) => (
+  <span
+    className="emoji"
+    role="img"
+    aria-label={props.label ? props.label : ""}
+    aria-hidden={props.label ? "false" : "true"}
+  >
+    {props.symbol}
+  </span>
+);
 export default Emoji;
