--- components/plannerPopup/transportForm.js
@@ -1,6 +1,0 @@
-import { Box } from "@chakra-ui/react";
-import React from "react";
-
-export default function TransportForm() {
-  return <Box>Hello Transport</Box>;
-}
