+++ components/sidebar.js
@@ -1,5 +1,7 @@
 import { Box, Icon } from "@chakra-ui/react";
-import { FaHome, FaPassport, FaLayerGroup } from "react-icons/fa";
+import { IoIosHome } from "react-icons/io";
+import { FaPassport } from "react-icons/fa";
+import { GrCatalog } from "react-icons/gr";
 import { MdDashboard } from "react-icons/md";
 import Link from "next/link";
 
@@ -7,7 +9,7 @@ const menu = [
   {
     name: "Home",
     path: "/dashboard",
-    icon: FaHome,
+    icon: IoIosHome,
   },
   {
     name: "Trips",
@@ -17,7 +19,8 @@ const menu = [
   {
     name: "Catalog",
     path: "/catalog",
-    icon: FaLayerGroup,
+    icon: GrCatalog,
+    // TODO: Change this
   },
   {
     name: "Trip Board",
@@ -28,16 +31,9 @@ const menu = [
 
 export default function Sidebar() {
   const MenuButton = ({ name, icon, path }) => (
-    <Box
-      display="flex"
-      alignItems="center"
-      justifyContent="center"
-      h={20}
-      cursor="pointer"
-      _hover={{
-        background: "gray.600",
-      }}
-    >
+    <Box display="flex" alignItems="center" justifyContent="center" h={20} cursor="pointer" _hover={{
+      background: "gray.600",
+    }}>
       <Box as={Link} href={path}>
         <Icon as={icon} w={30} h={30} color="gray.500" />
       </Box>
