+++ package.json
@@ -12,22 +12,16 @@
     "@emotion/react": "^11.0.0",
     "@emotion/styled": "^11.0.0",
     "@fullpage/react-fullpage": "^0.1.18",
-    "@material-ui/core": "^4.11.2",
-    "chakra-ui-autocomplete": "^1.2.3",
+    "@material-ui/core": "^4.11.1",
     "easy-peasy": "^4.0.1",
     "formik": "^2.2.5",
     "framer-motion": "^2.9.4",
-    "leaflet": "^1.7.1",
-    "leaflet-defaulticon-compatibility": "^0.1.1",
     "material-table": "^1.69.2",
-    "moment": "^2.29.1",
     "next": "10.0.0",
     "react": "17.0.1",
     "react-countup": "^4.3.3",
-    "react-datepicker": "^3.3.0",
     "react-dom": "17.0.1",
     "react-icons": "^3.11.0",
-    "react-leaflet": "^3.0.5",
     "yup": "^0.30.0"
   },
   "devDependencies": {
