+++ pages/dashboard.js
@@ -9,12 +9,10 @@ import {
   SimpleGrid,
   Link,
 } from "@chakra-ui/react";
-import { useRouter } from "next/router";
 import Emoji from "../components/emoji";
 import Layout from "../components/layout";
 
 export default function Dashboard() {
-  const router = useRouter();
   return (
     <Layout isLogin={true} title="Dashboard">
       <Flex
@@ -23,12 +21,7 @@ export default function Dashboard() {
         justifyContent="center"
         w="100%"
       >
-        <Button
-          colorScheme="pink"
-          variant="solid"
-          size="lg"
-          onClick={() => router.push("/trips/create")}
-        >
+        <Button colorScheme="pink" variant="solid" size="lg">
           Create a Trip
         </Button>
         <Flex w="100%" justifyContent="center" mt={20} gridGap={20}>
