--- sample/tripList.js
@@ -1,61 +1,0 @@
-export const sample = {
-  result: [
-    {
-      title: "Family Disneyland adventure",
-      description: "Budget Friendly trip to see the highlights of Disneyland ",
-      destination: "Florida",
-      days: "4 Nights",
-      travelers: 4,
-      cost: "$2,000 to $2,750",
-      keywords: "Florida, Disney, Family",
-    },
-    {
-      title: "Weekend wine testing in Virgina",
-      description:
-        "Exciting trips for couples to explore the mountains of Virginia and taste local wines",
-      destination: "Virginia",
-      days: "3 Nights",
-      travelers: 2,
-      cost: "$1,000 to $1,500",
-      keywords: "Wine, Virgina, Romance, Mountains",
-    },
-    {
-      title: "Family Disneyland adventure",
-      description: "Budget Friendly trip to see the highlights of Disneyland ",
-      destination: "Florida",
-      days: "4 Nights",
-      travelers: 4,
-      cost: "$2,000 to $2,750",
-      keywords: "Florida, Disney, Family",
-    },
-    {
-      title: "Weekend wine testing in Virgina",
-      description:
-        "Exciting trips for couples to explore the mountains of Virginia and taste local wines",
-      destination: "Virginia",
-      days: "3 Nights",
-      travelers: 2,
-      cost: "$1,000 to $1,500",
-      keywords: "Wine, Virgina, Romance, Mountains",
-    },
-    {
-      title: "Family Disneyland adventure",
-      description: "Budget Friendly trip to see the highlights of Disneyland ",
-      destination: "Florida",
-      days: "4 Nights",
-      travelers: 4,
-      cost: "$2,000 to $2,750",
-      keywords: "Florida, Disney, Family",
-    },
-    {
-      title: "Weekend wine testing in Virgina",
-      description:
-        "Exciting trips for couples to explore the mountains of Virginia and taste local wines",
-      destination: "Virginia",
-      days: "3 Nights",
-      travelers: 2,
-      cost: "$1,000 to $1,500",
-      keywords: "Wine, Virgina, Romance, Mountains",
-    },
-  ],
-};
