--- sample/tripPlan.js
@@ -1,24 +1,0 @@
-export const sample = {
-  result: [
-    {
-      id: 1,
-      title: "Day 1",
-      type: "day",
-      order: 1,
-      items: [
-        {
-          title: "2 nights accomodation at Taj but it was boring to see",
-          type: "lodging",
-        },
-        {
-          title: "Board flight to Paris",
-          type: "transport",
-        },
-        {
-          title: "Attend Sunburn at Goa",
-          type: "event",
-        },
-      ],
-    },
-  ],
-};
