--- services/trips.js
@@ -1,5 +1,0 @@
-import { sample as tripListJson } from "../sample/tripList";
-import { sample as tripPlanJson } from "../sample/tripPlan";
-
-export const tripList = () => tripListJson;
-export const tripPlan = () => tripPlanJson;
