+++ styles/global.css
@@ -1,5 +1,1 @@
 @import url('https://fonts.googleapis.com/css2?family=Kalam:wght@300;400;700&family=Kaushan+Script&family=Oxygen:wght@300;400;700&display=swap');
-
-button:focus{
-    outline:0 !important;
-}
